+++ src/helpers/uidMap.js
@@ -1,0 +1,3 @@
+const uidMap = {}
+
+export const createUid = () => Math.random().toString(36).substr(2, 9)